+++ containers/LoginForm/reducer.js
@@ -5,9 +5,6 @@ import { Validator } from './../../utils/validator';
 
 const validator = new Validator(inputsConfig);
 
-let initialInputChangePerformed = false;
-
-// TODO: check loginResultId, logOutResultId, cookieCheckResultId, changeId
 const defaultLoginState = Immutable.fromJS({
     authenticated: false,
     shouldChangePassword: false,
@@ -16,25 +13,48 @@ const defaultLoginState = Immutable.fromJS({
     loginForm: {
         inputs: getInputs(['username']),
         formError: '',
-        isFormValid: false,
-        invalidField: ''
-    }
+        shouldSubmit: false,
+        invalidField: '',
+        title: 'Login',
+        buttonLabel: 'Next'
+    },
+    loginData: {}
 });
+
+const updateLoginData = (state) => {
+    let inputs = state.getIn(['loginForm', 'inputs']);
+    let currentLoginData = state.get('loginData');
+
+    inputs.toSeq().forEach(input => {
+        if (!input.get('skipSubmit')) {
+            currentLoginData = currentLoginData.set(input.get('name'), input.get('value'));
+        }
+    });
+
+    return currentLoginData;
+};
 
 export const login = (state = defaultLoginState, action) => {
     let validationResult;
-
     switch (action.type) {
         case LOGOUT:
-            initialInputChangePerformed = false;
-            state = defaultLoginState;
             return state;
         case LOGIN:
             if (action.methodRequestState === 'finished') {
-                // TODO: change condition
+                state = state.setIn(['loginForm', 'shouldSubmit'], false);
+                // show password input and change title
                 if (action.error && action.error.type === 'policy.param.password') {
-                // merge rendered username input with the new password input
-                    return state.setIn(['loginForm', 'inputs', 'password'], Immutable.fromJS(getInputs(['password']).password));
+                    return state.setIn(['loginForm', 'inputs', 'password'], Immutable.fromJS(getInputs(['password']).password))
+                                .setIn(['loginForm', 'inputs', 'username', 'disabled'], true)
+                                .setIn(['loginForm', 'title'], Immutable.fromJS('Login with password'))
+                                .setIn(['loginForm', 'buttonLabel'], 'Login');
+                } else if (action.error && action.error.type === 'policy.param.newPassword') {
+                    // take only the username input from the current state and merge it with the new inputs for this step
+                    let newInputs = state.getIn(['loginForm', 'inputs']).take(1).merge(getInputs(['newPassword', 'confirmPassword']));
+
+                    return state.setIn(['loginForm', 'title'], 'Password change required')
+                                .setIn(['loginForm', 'inputs'], newInputs)
+                                .setIn(['loginForm', 'buttonLabel'], 'Change');
                 } else if (action.error) {
                     return state.setIn(['loginForm', 'formError'], action.error.message);
                 } else if (action.result) {
@@ -44,22 +64,24 @@ export const login = (state = defaultLoginState, action) => {
                                 .setIn(['loginForm', 'formError'], '');
                 }
             }
-
             return state;
         case SET_INPUT_VALUE:
             let { input, value } = action;
-            initialInputChangePerformed = true;
-
             return state.setIn(['loginForm', 'inputs', input, 'value'], value);
         case VALIDATE_FORM:
-            // submitAfter to detect if validate comes from blur or submit
-            validationResult = validator.validateAll(state.get('loginForm').get('inputs'));
+            let inputs = state.getIn(['loginForm', 'inputs']);
+            let loginData = state.get('loginData');
 
-            return initialInputChangePerformed ? state
-                .setIn(['loginForm', 'isFormValid'], validationResult.isValid)
-                .setIn(['loginForm', 'formError'], validationResult.error)
-                .setIn(['loginForm', 'invalidField'], validationResult.invalidField) : state;
+            validationResult = validator.validateAll(inputs);
 
+            if (validationResult.isValid) {
+                loginData = updateLoginData(state);
+            }
+
+            return state.set('loginData', loginData)
+                        .setIn(['loginForm', 'shouldSubmit'], validationResult.isValid)
+                        .setIn(['loginForm', 'formError'], validationResult.error)
+                        .setIn(['loginForm', 'invalidField'], validationResult.invalidField);
         case COOKIE_CHECK:
             if (action.methodRequestState === 'finished') {
                 if (action.error) {